* Initial commit!
Note that the repository contents are based on the Canonical Debate Lab client repo (https://github.com/canonical-debate-lab/client), which was based on my original Debate Map repo (https://github.com/Venryx/DebateMap).

The reason I'm rebranding the repo under "Debate Map" instead of "Canonical Debate Lab" is that the group has decided to use a "two-tiered" development approach:
Tier1) A shared "CD service layer" that holds the data found useful for all/many of our projects, eg. open claims and arguments. It will also probably include a "base client", that is not as fully-featured, but developed enough that browsing the CD layer's data is possible. (albeit, probably not as efficiently)
Tier2) The custom clients (and their databases, tools, etc.), specific to the vision that each member of the group has. For me, that's Debate Map. For Tim, that's Gruff. For Jamie, that's SL/GAD. And so forth. We use the CD service layer for the "agreed upon" shared data, but augment that with our own tier-2 stuff for all the features we want that are too specific/opinionated, or that we haven't agreed upon (yet anyway).

Onward!